Reset react native files
