+++ frontend/app-example/(tabs)/explore.tsx
